+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project are documented in this file.
  - Dark theme for the Fluent style
  - Added `fluent-light` and `fluent-dark` as explicit styles to select a light/dark variant,
    regardless of the system color scheme setting.
+ - Added `Window::is_visible` 
 
 ### Fixed